Define MaxUpdatesPerCycle for updates on 2008 R2 builds
Previously, the 2008 R2 build would time out due to an unidentified
problem during the update process. It seems that limiting the number
of updates applied at once solves the problem.

Until the problem is identified and a proper fix formed, the
Autounattend scripts for 2008 R2 has been updated to make use of this
workaround. The problem persists with 2008 R2 Core even with this
addition.