Fix syntax that prevents update discovery
The Count and other properties of the items were not accessible if there
is only 1 item in the array. This fix ensures that all properties can be
read.

This was tested and discovered in Windows Server 2012 R2. This results
in a longer update time, which proves that there are now more updates
being installed.